+++ cli/engine/PrepareEngine.py
@@ -4,7 +4,7 @@ import stat
 from cli.helpers.Step import Step
 from cli.helpers.data_loader import BASE_DIR
 from cli.helpers.Config import Config
-from cli.helpers.build_saver import copy_files_recursively
+from cli.helpers.build_io import copy_files_recursively
 
 
 class PrepareEngine(Step):
