+++ cli/engine/TestEngine.py
@@ -1,7 +1,7 @@
 import os
 
 from cli.helpers.Step import Step
-from cli.helpers.build_saver import SPEC_OUTPUT_DIR, ANSIBLE_INVENTORY_FILE
+from cli.helpers.build_io import SPEC_OUTPUT_DIR, ANSIBLE_INVENTORY_FILE
 from cli.helpers.doc_list_helpers import select_single
 from cli.engine.spec.SpecCommand import SpecCommand
 from cli.helpers.data_loader import load_manifest
