+++ cli/engine/ansible/AnsibleInventoryUpgrade.py
@@ -2,10 +2,10 @@ from ansible.parsing.dataloader import DataLoader
 from ansible.inventory.manager import InventoryManager
 
 from cli.helpers.Step import Step
-from cli.helpers.build_saver import get_inventory_path_for_build
+from cli.helpers.build_io import get_inventory_path_for_build
 from cli.models.AnsibleHostModel import AnsibleHostModel
 from cli.models.AnsibleInventoryItem import AnsibleInventoryItem
-from cli.helpers.build_saver import save_inventory
+from cli.helpers.build_io import save_inventory
 from cli.helpers.objdict_helpers import dict_to_objdict
 from cli.helpers.data_loader import load_schema_obj, types as data_types
 from cli.helpers.doc_list_helpers import select_single
