+++ cli/engine/providers/aws/InfrastructureBuilder.py
@@ -8,7 +8,7 @@ from cli.helpers.config_merger import merge_with_defaults
 from cli.engine.providers.aws.APIProxy import APIProxy
 from cli.helpers.Step import Step
 from cli.helpers.doc_list_helpers import select_single, select_all
-from cli.helpers.build_saver import get_terraform_path
+from cli.helpers.build_io import get_terraform_path
 from cli.helpers.data_loader import load_json_obj
 from cli.helpers.naming_helpers import resource_name
 from cli.helpers.objdict_helpers import objdict_to_dict, dict_to_objdict
