+++ conftest.py
@@ -2,7 +2,7 @@ import json
 import os
 import shutil
 
-from cli.helpers.build_saver import save_manifest, MANIFEST_FILE_NAME
+from cli.helpers.build_io import save_manifest, MANIFEST_FILE_NAME
 from cli.helpers.Config import Config
 from tests.unit.helpers.constants import TEST_DOCS, CLUSTER_NAME_SAVE, CLUSTER_NAME_LOAD, NON_EXISTING_CLUSTER,\
     TEST_JSON, TEST_JSON_NAME
