+++ tests/unit/helpers/test_build_io.py
@@ -3,16 +3,17 @@ import os
 
 from ruamel.yaml import YAML
 
-from cli.helpers.build_saver import get_build_path, get_output_path, get_terraform_path, get_ansible_path,\
+from cli.helpers.build_io import get_build_path, get_output_path, get_terraform_path, get_ansible_path,\
     get_ansible_vault_path, get_ansible_config_file_path, get_inventory_path, get_manifest_path,\
-    save_manifest, save_sp, save_inventory, save_ansible_config_file, get_inventory_path_for_build,\
+    save_manifest, load_manifest, save_sp, save_inventory, save_ansible_config_file, get_inventory_path_for_build,\
     get_ansible_config_file_path_for_build, get_ansible_path_for_build,\
     ANSIBLE_OUTPUT_DIR, ANSIBLE_VAULT_OUTPUT_DIR, ANSIBLE_INVENTORY_FILE, ANSIBLE_CFG_FILE,\
     MANIFEST_FILE_NAME, SP_FILE_NAME, TERRAFORM_OUTPUT_DIR
 from cli.helpers.objdict_helpers import dict_to_objdict
 from cli.helpers.yaml_helpers import safe_load_all, safe_load
 
-from tests.unit.helpers.constants import CLUSTER_NAME_SAVE, OUTPUT_PATH, TEST_DOCS, TEST_CLUSTER_MODEL, TEST_INVENTORY
+from tests.unit.helpers.constants import CLUSTER_NAME_SAVE, OUTPUT_PATH, TEST_DOCS, TEST_CLUSTER_MODEL, \
+    TEST_INVENTORY, CLUSTER_NAME_LOAD
 
 TEST_SP = {'appId': 'xxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxx',
            'displayName': 'test-rg',
@@ -98,6 +99,12 @@ def test_save_manifest():
     manifest_stream = open(manifest_path, 'r')
     manifest_file_content = safe_load_all(manifest_stream)
     assert TEST_DOCS == manifest_file_content
+
+
+def test_load_manifest():
+    build_path = get_build_path(CLUSTER_NAME_LOAD)
+    docs = load_manifest(build_path)
+    assert docs == TEST_DOCS
 
 
 def test_save_sp():
